Move upload functionality to lib/projects/upload
