+++ lang/en.lyaml
@@ -1094,16 +1094,6 @@ en:
       checkIfParentAccountIsAuthed:
         notAuthedError: "To develop this project locally, run {{ authCommand }} to authenticate the App Developer Account {{ accountId }} associated with {{ accountIdentifier }}."
     projects:
-      uploadProjectFiles:
-        add: "Uploading {{#bold}}{{ projectName }}{{/bold}} project files to {{ accountIdentifier }}"
-        fail: "Failed to upload {{#bold}}{{ projectName }}{{/bold}} project files to {{ accountIdentifier }}"
-        succeed: "Uploaded {{#bold}}{{ projectName }}{{/bold}} project files to {{ accountIdentifier }}"
-        buildCreated: "Project \"{{ projectName }}\" uploaded and build #{{ buildId }} created"
-      handleProjectUpload:
-        emptySource: "Source directory \"{{ srcDir }}\" is empty. Add files to your project and rerun `{{#yellow}}hs project upload{{/yellow}}` to upload them to HubSpot."
-        compressed: "Project files compressed: {{ byteCount }} bytes"
-        compressing: "Compressing build files to \"{{ path }}\""
-        fileFiltered: "Ignore rule triggered for \"{{ filename }}\""
       validateProjectConfig:
         configNotFound: "Unable to locate a project configuration file. Try running again from a project directory, or run {{ createCommand }} to create a new project."
         configMissingFields: "The project configuruation file is missing required fields."
@@ -1132,6 +1122,17 @@ en:
         buildSucceededAutomaticallyDeploying: "Build #{{ buildId }} succeeded. {{#bold}}Automatically deploying{{/bold}} to {{ accountIdentifier }}\n"
         cleanedUpTempFile: "Cleaned up temporary file {{ path }}"
         viewDeploys: "View all deploys for this project in HubSpot"
+    projectUpload:
+      uploadProjectFiles:
+        add: "Uploading {{#bold}}{{ projectName }}{{/bold}} project files to {{ accountIdentifier }}"
+        fail: "Failed to upload {{#bold}}{{ projectName }}{{/bold}} project files to {{ accountIdentifier }}"
+        succeed: "Uploaded {{#bold}}{{ projectName }}{{/bold}} project files to {{ accountIdentifier }}"
+        buildCreated: "Project \"{{ projectName }}\" uploaded and build #{{ buildId }} created"
+      handleProjectUpload:
+        emptySource: "Source directory \"{{ srcDir }}\" is empty. Add files to your project and rerun `{{#yellow}}hs project upload{{/yellow}}` to upload them to HubSpot."
+        compressed: "Project files compressed: {{ byteCount }} bytes"
+        compressing: "Compressing build files to \"{{ path }}\""
+        fileFiltered: "Ignore rule triggered for \"{{ filename }}\""
     ui:
       betaTag: "{{#bold}}[BETA]{{/bold}}"
       betaWarning:
