+++ lib/projects/index.ts
@@ -1,17 +1,13 @@
 import fs from 'fs-extra';
 import path from 'path';
-import archiver from 'archiver';
-import tmp, { FileResult } from 'tmp';
 import findup from 'findup-sync';
 import { logger } from '@hubspot/local-dev-lib/logger';
 import { fetchFileFromRepository } from '@hubspot/local-dev-lib/github';
 import {
   createProject,
   fetchProject,
-  uploadProject,
 } from '@hubspot/local-dev-lib/api/projects';
 import { isSpecifiedError } from '@hubspot/local-dev-lib/errors/index';
-import { shouldIgnoreFile } from '@hubspot/local-dev-lib/ignoreRules';
 import { getCwd, getAbsoluteFilePath } from '@hubspot/local-dev-lib/path';
 import { downloadGithubRepoContents } from '@hubspot/local-dev-lib/github';
 import { RepoPath } from '@hubspot/local-dev-lib/types/Github';
@@ -37,7 +33,7 @@ import {
   ProjectAddComponentData,
   ProjectTemplateRepoConfig,
   ComponentTemplate,
-} from '../../types/projects';
+} from '../../types/Projects';
 
 const i18nKey = 'lib.projects';
 
@@ -322,169 +318,6 @@ export async function ensureProjectExists(
     logError(err, new ApiErrorContext({ accountId }));
     process.exit(EXIT_CODES.ERROR);
   }
-}
-
-async function uploadProjectFiles(
-  accountId: number,
-  projectName: string,
-  filePath: string,
-  uploadMessage: string,
-  platformVersion: string
-): Promise<{ buildId?: number; error: unknown }> {
-  SpinniesManager.init({});
-  const accountIdentifier = uiAccountDescription(accountId);
-
-  SpinniesManager.add('upload', {
-    text: i18n(`${i18nKey}.uploadProjectFiles.add`, {
-      accountIdentifier,
-      projectName,
-    }),
-    succeedColor: 'white',
-  });
-
-  let buildId: number | undefined;
-  let error: unknown;
-
-  try {
-    const { data: upload } = await uploadProject(
-      accountId,
-      projectName,
-      filePath,
-      uploadMessage,
-      platformVersion
-    );
-
-    buildId = upload.buildId;
-
-    SpinniesManager.succeed('upload', {
-      text: i18n(`${i18nKey}.uploadProjectFiles.succeed`, {
-        accountIdentifier,
-        projectName,
-      }),
-    });
-
-    logger.debug(
-      i18n(`${i18nKey}.uploadProjectFiles.buildCreated`, {
-        buildId,
-        projectName,
-      })
-    );
-  } catch (err) {
-    SpinniesManager.fail('upload', {
-      text: i18n(`${i18nKey}.uploadProjectFiles.fail`, {
-        accountIdentifier,
-        projectName,
-      }),
-    });
-
-    error = err;
-  }
-
-  return { buildId, error };
-}
-
-type ProjectUploadCallbackFunction<T> = (
-  accountId: number,
-  projectConfig: ProjectConfig,
-  tempFile: FileResult,
-  buildId?: number
-) => Promise<T | undefined>;
-
-type ProjectUploadDefaultResult = {
-  uploadError?: unknown;
-};
-
-export async function handleProjectUpload<T = ProjectUploadDefaultResult>(
-  accountId: number,
-  projectConfig: ProjectConfig,
-  projectDir: string,
-  callbackFunc: ProjectUploadCallbackFunction<T>,
-  uploadMessage: string
-) {
-  const srcDir = path.resolve(projectDir, projectConfig.srcDir);
-
-  const filenames = fs.readdirSync(srcDir);
-  if (!filenames || filenames.length === 0) {
-    logger.log(
-      i18n(`${i18nKey}.handleProjectUpload.emptySource`, {
-        srcDir: projectConfig.srcDir,
-      })
-    );
-    process.exit(EXIT_CODES.SUCCESS);
-  }
-
-  const tempFile = tmp.fileSync({ postfix: '.zip' });
-
-  logger.debug(
-    i18n(`${i18nKey}.handleProjectUpload.compressing`, {
-      path: tempFile.name,
-    })
-  );
-
-  const output = fs.createWriteStream(tempFile.name);
-  const archive = archiver('zip');
-
-  const result = new Promise(resolve =>
-    output.on('close', async function () {
-      let uploadResult: ProjectUploadDefaultResult | T | undefined;
-
-      logger.debug(
-        i18n(`${i18nKey}.handleProjectUpload.compressed`, {
-          byteCount: archive.pointer(),
-        })
-      );
-
-      const { buildId, error } = await uploadProjectFiles(
-        accountId,
-        projectConfig.name,
-        tempFile.name,
-        uploadMessage,
-        projectConfig.platformVersion
-      );
-
-      if (error) {
-        console.log(error);
-        uploadResult = { uploadError: error };
-      } else if (callbackFunc) {
-        console.log('callbackfunc');
-        uploadResult = await callbackFunc(
-          accountId,
-          projectConfig,
-          tempFile,
-          buildId
-        );
-      }
-      resolve(uploadResult || {});
-    })
-  );
-
-  archive.pipe(output);
-
-  let loggedIgnoredNodeModule = false;
-
-  archive.directory(srcDir, false, file => {
-    const ignored = shouldIgnoreFile(file.name, true);
-    if (ignored) {
-      const isNodeModule = file.name.includes('node_modules');
-
-      if (!isNodeModule || !loggedIgnoredNodeModule) {
-        logger.debug(
-          i18n(`${i18nKey}.handleProjectUpload.fileFiltered`, {
-            filename: file.name,
-          })
-        );
-      }
-
-      if (isNodeModule && !loggedIgnoredNodeModule) {
-        loggedIgnoredNodeModule = true;
-      }
-    }
-    return ignored ? false : file;
-  });
-
-  archive.finalize();
-
-  return result;
 }
 
 export function logFeedbackMessage(buildId: number): void {
