+++ lib/prompts/sandboxesPrompt.ts
@@ -9,7 +9,7 @@ import {
   getConfigAccounts,
 } from '@hubspot/local-dev-lib/config';
 import { CLIAccount } from '@hubspot/local-dev-lib/types/Accounts';
-import { PromptChoices } from '../../types/prompts';
+import { PromptChoices } from '../../types/Prompts';
 
 const i18nKey = 'lib.prompts.sandboxesPrompt';
